결제관련 데이터타입 BigDecimal 반영 (#70)
* Create dfsdf.txt

* pass 도메인 서비스/엔티티/레포지토리 1차 구현

* Create dfsdf - 복사본.txt

* pass -api 설정

* api 주소 설정

* test

* 서버동작

* 유저 테이블 정합성 및 빌드 오류 수정

스웨거까지

* 스웨거 구현+페이먼트정리

* 머지후 에러수정

* Update fitneedsdb.sql

* 기능정리

* pass-layout

* 이용권거래 프론트 적용

* 머지 페이먼트 삭제

* pass-payment

* api연동

* userpass merge

* userpass/payment 병합

* 유저패스/페이먼트 병합완료

서버정상구동완료

* Update UserPassTradeService.java

* passtradeSQL

insert SAMPLE upload

* Update PassTradeSql.sql

* 불필요파일삭제

박창일

* db - 경로

박창일

* 가데이터 test

박창일

* 내 이용권 조회 API DTO 분리 및 fetch join 적용 / 가데이터 조회 테스트완료

박창일
[UserPass] 이용권 조회 API DTO 분리

[Pass] 내 이용권 조회 fetch join 적용

* sql 수정

박창일

* sql 수정

박창일

* Update pass_trade.sql

* sql 정리

박창일
pass sql 정리

* Revert "sql 정리"

This reverts commit d828261505e245fe05879411b9e35968ca7b0a42.

* Reapply "sql 정리"

This reverts commit 5a6fb9a9b6cece121a10ab82fdb5ebb69849ccc2.

* deve260101 머지반영

박창일

* 등록게시글 수정/삭제 로직 추가

* ui 모달테스트

* dev260101 머지 / 금액관련 데이터타입 수정

---------

+++ enduser-backend/src/main/java/com/project/app/pass_trade/dto/response/PassTradeTransactionResponse.java
@@ -2,8 +2,9 @@ package com.project.app.pass_trade.dto.response;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
-import com.project.app.pass_trade.domain.TransactionStatus;
+import com.project.app.pass_trade.entity.TransactionStatus;
 
 
 @Data
@@ -13,7 +14,7 @@ public class PassTradeTransactionResponse {
     private String buyerId;
     private String sellerId;
     private Integer tradeCount;
-    private Integer totalAmount;
+    private BigDecimal totalAmount;
     private TransactionStatus transactionStatus;
     private Long paymentId;
     private LocalDateTime regDt;
